unsigned long long

+++ insertion/cpp/insertionNomes.cpp
@@ -10,7 +10,7 @@ int main(int argc, char *argv[2])
     std::ofstream output("./insCppNomes.txt", std::ios_base::app);
     std::vector<std::string> vetor;
     std::string aux;
-    unsigned long comp = 0, swap = 0;
+    unsigned long long comp = 0, swap = 0;
     clock_t inicio, fim;
 
     for (int i = 0; i < atoi(argv[1]); i++)
